model/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI & LLM - Compatible versions
langchain-groq==0.1.9
langchain-core==0.2.38
langchain==0.2.16
pydantic==2.5.3

# PDF Processing
PyMuPDF==1.23.8

# Configuration
python-dotenv==1.0.0

# Utilities
aiofiles==23.2.1
requests==2.31.0